💄 Add google fonts to plugin
