+++ packages/starknet_flutter/lib/src/views/add_another_wallet/widgets/wallet_action_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:starknet_flutter/src/views/widgets/bouncing_button.dart';
+import 'package:starknet_flutter/src/views/widgets/bouncing_widget.dart';
 
 class WalletActionButton extends StatelessWidget {
   final String title;
