+++ packages/starknet_flutter/lib/src/views/wallet/wallet_initialization.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:starknet_flutter/src/views/utils/snackbar_utils.dart';
 import 'package:starknet_flutter/src/views/wallet/routes/welcome/wallet_welcome_view.dart';
 import 'package:starknet_flutter/src/views/wallet/wallet_initialization_observer.dart';
 import 'package:starknet_flutter/src/views/wallet/wallet_initialization_router.dart';
 import 'package:starknet_flutter/src/views/wallet_list/wallet_list_viewmodel.dart';
-import 'package:starknet_flutter/src/views/widgets/bouncing_button.dart';
+import 'package:starknet_flutter/src/views/widgets/bouncing_widget.dart';
 
 import '../passcode/passcode_input_view.dart';
 import 'wallet_initialization_presenter.dart';
@@ -93,47 +94,51 @@ class _WalletInitializationPageState extends State<WalletInitializationPage>
 
   @override
   Widget build(BuildContext modalContext) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        elevation: 0,
-        title: model.title != null
-            ? Text(
-                model.title!,
-                style: TextStyle(
-                  color: Theme.of(context).colorScheme.onSurface,
-                  fontSize: 16,
-                ),
-              )
-            : const SizedBox.shrink(),
-        leading: AnimatedSwitcher(
-          duration: const Duration(milliseconds: 300),
-          child: isRootPage
-              ? const SizedBox.shrink()
-              : BouncingWidget(
-                  onTap: goBack,
-                  child: Container(
-                    color: Colors.transparent,
-                    width: 50,
-                    height: 50,
-                    child: Icon(
-                      Icons.arrow_back_ios_new_rounded,
-                      color: Theme.of(context).primaryColor,
+    return GestureDetector(
+      onTap: () => FocusScope.of(context).unfocus(),
+      child: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.transparent,
+          elevation: 0,
+          title: model.title != null
+              ? Text(
+                  model.title!,
+                  style: GoogleFonts.poppins(
+                    color: Theme.of(context).colorScheme.onSurface,
+                    fontSize: 16,
+                    fontWeight: FontWeight.w700,
+                  ),
+                )
+              : const SizedBox.shrink(),
+          leading: AnimatedSwitcher(
+            duration: const Duration(milliseconds: 300),
+            child: isRootPage
+                ? const SizedBox.shrink()
+                : BouncingWidget(
+                    onTap: goBack,
+                    child: Container(
+                      color: Colors.transparent,
+                      width: 50,
+                      height: 50,
+                      child: Icon(
+                        Icons.arrow_back_ios_new_rounded,
+                        color: Theme.of(context).primaryColor,
+                      ),
                     ),
                   ),
-                ),
+          ),
         ),
-      ),
-      body: SafeArea(
-        child: Navigator(
-          key: _navigatorKey,
-          initialRoute: widget.initialRoute,
-          observers: [observer],
-          onGenerateRoute: (settings) =>
-              WalletInitializationRouter.onGenerateRoute(
-            settings,
-            model,
-            presenter,
+        body: SafeArea(
+          child: Navigator(
+            key: _navigatorKey,
+            initialRoute: widget.initialRoute,
+            observers: [observer],
+            onGenerateRoute: (settings) =>
+                WalletInitializationRouter.onGenerateRoute(
+              settings,
+              model,
+              presenter,
+            ),
           ),
         ),
       ),
