+++ packages/starknet_flutter/lib/src/views/wallet_list/wallet_list.dart
@@ -1,12 +1,13 @@
 import 'dart:core';
 
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:starknet_flutter/src/views/add_another_wallet/add_another_wallet.dart';
 import 'package:starknet_flutter/src/views/views.dart';
 import 'package:starknet_flutter/src/views/wallet_list/widgets/appbar.dart';
 import 'package:starknet_flutter/src/views/wallet_list/widgets/wallet_cell.dart';
-import 'package:starknet_flutter/src/views/widgets/starknet_button.dart';
+import 'package:starknet_flutter/src/views/widgets/bouncing_button.dart';
 
 import 'wallet_list_presenter.dart';
 import 'wallet_list_viewmodel.dart';
@@ -110,11 +111,11 @@ class _WalletListPageState extends State<WalletListPage>
                                     },
                                     itemCount: wallets.length,
                                   )
-                                : const Padding(
-                                    padding: EdgeInsets.symmetric(vertical: 30),
+                                : Padding(
+                                    padding: const EdgeInsets.symmetric(vertical: 30),
                                     child: Text(
                                       'No wallets added yet',
-                                      style: TextStyle(
+                                      style: GoogleFonts.poppins(
                                         color: Colors.grey,
                                         fontWeight: FontWeight.bold,
                                       ),
@@ -124,12 +125,12 @@ class _WalletListPageState extends State<WalletListPage>
                         }),
                   ),
                   const SizedBox(height: 5),
-                  StarknetButton.text(
+                  BouncingButton.text(
                     onTap: () {
                       closeModal(() => openAddAnotherWalletModal());
                     },
                     text: 'Add another wallet',
-                    textStyle: TextStyle(
+                    textStyle: GoogleFonts.poppins(
                       color: Theme.of(context).primaryColor,
                       fontWeight: FontWeight.bold,
                     ),
